Fix integer overflow in StringUtil::implode
Reviewed By: ricklavoie

Differential Revision: D3623922

fbshipit-source-id: 136d124a850c07cc6c63535afc11d36499d576fc

+++ hphp/runtime/base/string-util.cpp
@@ -123,9 +123,9 @@ String StringUtil::Implode(const Variant& items, const String& delim,
   if (size == 0) return empty_string();
 
   String* sitems = (String*)req::malloc(size * sizeof(String));
-  int len = 0;
-  int lenDelim = delim.size();
-  int i = 0;
+  size_t len = 0;
+  size_t lenDelim = delim.size();
+  size_t i = 0;
   for (ArrayIter iter(items); iter; ++iter) {
     new (&sitems[i]) String(iter.second().toString());
     len += sitems[i].size() + lenDelim;
